pyproject.toml
[project]
name = "beework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "lmnr>=0.7.41",
    "modal>=1.3.3",
    "python-dotenv>=1.2.1",
]
